update resource view all

+++ app/Http/Controllers/MaterialResource.php
@@ -17,7 +17,7 @@ class MaterialResource extends Controller
      */
     public function index()
     {
-        $materials = Material::paginate(2);
+        $materials = Material::paginate(10);
         $materialCategory=MaterialCategory::whereNotIn('id',[998])->get();
         return view('material.indexMaterial', compact('materials','materialCategory'));
     }
@@ -132,7 +132,7 @@ class MaterialResource extends Controller
         ];
 
         if($request->hasFile('material_image')){
-            if($material->material_image == '' || $material->material_image == null){
+            if(!$material->material_image == '' && !$material->material_image == null){
                 unlink('uploads/material/'.$material->material_image);
             }
             $file = $request->file('material_image');
@@ -154,8 +154,7 @@ class MaterialResource extends Controller
      */
     public function destroy(Material $material)
     {
-
-        if($material->material_image != 'default.jpg'){
+        if($material->material_image != 'default.jpg' && !$material->material_image == '' && !$material->material_image == null){
             unlink('uploads/material/'.$material->material_image);
         }
         $material->delete();
